requirements_backend.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
Flask-Limiter==3.5.0
redis==5.0.1
cryptography==41.0.7
gunicorn==21.2.0
ijson==3.2.3
supabase==2.6.0
